requirements.txt
aniso8601==9.0.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
mysqlclient==2.0.3
PyMySQL==1.0.2
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
